worker/requirements.txt
# networking service
uvloop==0.17.0
msgspec==0.18.2
cloudpickle==2.2.1
pyzmq==25.1.1
aiozmq==1.0.0
# monitoring service
# psutil==5.9.2
# Kafka
aiokafka==0.8.1
# snapshots
minio==7.1.16
# incremental snapshots
git+https://github.com/delftdata/stateflow-kevo.git
# conflict detection
networkx==3.1

# deathstar
geopy
